component CSS and files cleanup
