+++ src/About.scss
@@ -1,59 +1,37 @@
 @import "./UniversalStyles";
 
 .about {
-  width: 960px;
-  margin: 0 auto;
-  margin-top: 48px;
-  margin-bottom: 48px;
-
-  @media (max-width: 1023px) {
-    padding: 0 32px;
-    box-sizing: border-box;
-    width: 100%;
-  }
-
+  padding-top: 3%;
   h1 {
     color: $color2;
-    font-size: 60px;
     font-weight: 500;
     line-height: 72px;
-
-    @media (max-width: 639px) {
-      font-size: 60px * 0.75;
-      line-height: 72px * 0.75;
-    }
   }
 
   h2 {
     color: $color1;
-    font-size: 36px;
     font-weight: 400;
     line-height: 44px;
     margin-top: 24px;
     margin-bottom: 24px;
+  }
 
-    @media (max-width: 639px) {
-      font-size: 36px * 0.75;
-      line-height: 44px * 0.75;
+  @include media-breakpoint-down(sm) { 
+    h1, h2 {
+      line-height: 1;
     }
   }
 
   .text {
     color: #000;
-    font-size: 20px;
+    font-size: 1.2rem;
     line-height: 28px;
     font-weight: 300;
     margin-top: 16px;
 
-    @media (max-width: 639px) {
-      font-size: 20px * 0.75;
-      line-height: 28px * 0.75;
-      margin-top: 12px;
-    }
-
     & > .people {
       color: $color2;
-      font-size: 18px;
+      font-size: 1.1rem;
       line-height: 24px;
       font-weight: 400;
       display: block;
@@ -62,9 +40,7 @@
       &:nth-of-type(1) {
         margin-top: 12px;
       }
-
-      @media (max-width: 639px) {
-        font-size: 18px * 0.75;
+      @include media-breakpoint-only(sm) { 
         line-height: 24px * 0.75;
         margin-top: 4px;
 
@@ -84,11 +60,6 @@
     margin-top: 24px;
     margin-bottom: 48px;
 
-    @media (max-width: 639px) {
-      margin-top: 18px;
-      margin-bottom: 32px;
-    }
-
     .content {
       opacity: 1;
     }
@@ -101,13 +72,13 @@
       padding: 16px 48px;
       padding-bottom: 0;
 
-      @media (max-width: 639px) {
-        padding: 8px 32px;
+      @include media-breakpoint-only(sm) {
+        padding: 8px 16px;
       }
 
       .name {
         color: $color2;
-        font-size: 18px;
+        font-size: 1.1rem;
         line-height: 24px;
         font-weight: 400;
         display: block;
@@ -115,21 +86,19 @@
         margin-bottom: 1px;
 
         @media (max-width: 639px) {
-          font-size: 18px * 0.75;
           line-height: 24px * 0.75;
         }
       }
 
       .department {
         color: #666;
-        font-size: 17px;
+        font-size: 1rem;
         line-height: 21px;
         font-weight: 300;
         display: block;
         padding-bottom: 16px;
 
-        @media (max-width: 639px) {
-          font-size: 17px * 0.75;
+        @include media-breakpoint-only(sm) {
           line-height: 21px * 0.75;
           padding-top: 8px;
         }
@@ -145,7 +114,7 @@
 img.article-image {
   width: 100%;
   margin-top: 24px;
-  @media (max-width: 639px) {
+  @include media-breakpoint-only(sm) {
     width: 100%;
     margin-top: 16px;
   }
