+++ src/UniversalStyles.scss
@@ -2,23 +2,3 @@
 @import '../node_modules/bootstrap/scss/mixins/_breakpoints';
 $color1: #3f51b2;
 $color2: #ff6f00;
-
-.pr-0 {
-	padding-right: 0;
-}
-
-.pl-0 {
-  padding-left: 0;
-}
-
-.pt-7 {
-  padding-top: 7px;
-}
-
-.pt-14 {
-  padding-top: 14px;
-}
-
-.pb-14 {
-  padding-bottom: 14px;
-}
