+++ src/components/AccordionTab/AccordionTab.scss
@@ -1,4 +1,4 @@
-@import "../UniversalStyles";
+@import "./UniversalStyles";
 
 @mixin tab {
   border-bottom: 1px solid #ccc;
@@ -13,15 +13,14 @@
 
   .title {
     color: $color2;
-    font-size: 24px;
+    font-size: 1.2rem;
     font-weight: 500;
     cursor: pointer;
     width: 100%;
     padding: 24px;
     margin: -24px;
 
-    @media (max-width: 639px) {
-      font-size: 24px * 0.75;
+    @include media-breakpoint-down(sm) { 
       padding: 16px;
       margin: -16px;
     }
@@ -29,7 +28,7 @@
 
   &:after {
     content: "";
-    background-image: url("../img/pp-map-arrow-icon-grey-down.png");
+    background-image: url("../../img/pp-map-arrow-icon-grey-down.png");
     background-repeat: no-repeat;
     width: 32px;
     height: 16px;
@@ -40,7 +39,7 @@
     top: 28px;
     right: 24px;
 
-    @media (max-width: 639px) {
+    @include media-breakpoint-down(sm) { 
       width: 24px;
       height: 12px;
       top: 17px;
@@ -61,7 +60,7 @@
   @include tab;
 
   &:after {
-    background-image: url("../img/pp-map-arrow-icon-grey-up.png");
+    background-image: url("../../img/pp-map-arrow-icon-grey-up.png");
   }
 
   div:nth-child(2) {
