+++ src/components/CaseEditor/CaseEditor.scss
@@ -1,5 +1,5 @@
 .addCaseForm {
-  padding: "3em";
+  padding: 3em;
 }
 
 .addCaseForm h4 {
