+++ src/components/MapVisualization/MapVisualization.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import { Map, TileLayer } from 'react-leaflet'
-import geojson from '../world-countries.json'
+import geojson from '../../world-countries.json'
 import Choropleth from 'react-leaflet-choropleth'
 import './MapVisualization.css'
-import api from '../utils/api'
-import caseMarkerIcon from '../img/case-marker-icon.png'
-import methodMarkerIcon from '../img/method-marker-icon.png'
-import orgMarkerIcon from '../img/organization-marker-icon.png'
-import mapArrowIcon from '../img/pp-map-arrow-icon.png'
+import api from '../../utils/api'
+import caseMarkerIcon from '../../img/case-marker-icon.png'
+import methodMarkerIcon from '../../img/method-marker-icon.png'
+import orgMarkerIcon from '../../img/organization-marker-icon.png'
+import mapArrowIcon from '../../img/pp-map-arrow-icon.png'
 
 import 'leaflet/dist/leaflet.css'
