+++ src/components/QuickSubmit/QuickSubmit.scss
@@ -1,4 +1,4 @@
-@import "../UniversalStyles";
+@import "../../UniversalStyles";
 
 .quick-submit {
   padding-top: 7%;
