+++ src/components/QuickSubmitPicker/QuickSubmitPicker.scss
@@ -1,4 +1,4 @@
-@import "../UniversalStyles";
+@import "../../UniversalStyles";
 
 .select-type {
   padding-top: 7%;
