+++ src/components/SearchQueryField/SearchQueryField.js
