+++ src/components/SearchResultsView/SearchResultsView.scss
@@ -21,7 +21,7 @@
       margin-top: 70px;
 
       a {
-        font-size: 16px;
+        font-size: 1rem;
         font-weight: 400;
         color: #666;
         display: block;
@@ -35,7 +35,7 @@
       }
 
       .current-sorting-selection {
-        font-size: 16px;
+        font-size: 1rem;
         font-weight: 400;
         color: $color2;
         display: block;
@@ -58,7 +58,7 @@
     }
 
     .filters a {
-      font-size: 16px;
+      font-size: 1rem;
       font-weight: 400;
       color: #000;
       display: block;
@@ -70,7 +70,7 @@
       border-bottom: 4px solid #fff;
     }
 
-    @media (max-width: 639px) {
+    @include media-breakpoint-down(sm) {
       .filters a {
         float: none;
         border: 0 !important;
@@ -124,7 +124,7 @@
 
   .results-box {
     .category-title {
-      font-size: 30px;
+      font-size: 1.8rem;
       font-weight: 400;
       color: #ff6f00;
       line-height: 30px;
@@ -141,7 +141,7 @@
   padding: 14px 0;
 
   p {
-    font-size: 16px;
+    font-size: 1rem;
     font-weight: 400;
     color: #999;
     line-height: 26px;
