+++ src/containers/Case/Case.scss
@@ -214,7 +214,6 @@
 
       .case-html {
         h3 {
-          font-size: 32px;
           font-weight: 400;
           line-height: 40px;
           margin-bottom: 12px;
@@ -222,7 +221,7 @@
         }
 
         p, li {
-          font-size: 16px;
+          font-size: 1rem;
           font-weight: 100;
           line-height: 24px;
           margin-bottom: 16px;
